# webapi/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "TakeMe",
    "description": "API documentation for TakeMe API"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/email/{username}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Retrieve email based on a given username",
        "description": "",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "lefan@gmail.com"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/{username}/page": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User data (without page) through JWT",
        "description": "",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User data (without page) through JWT",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update TakeMe user data through JWT",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/users/create": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new user. Returns 404 if user already exists",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "lefanTan"
                },
                "email": {
                  "type": "string",
                  "example": "lefantan@lol.com"
                },
                "password": {
                  "type": "string",
                  "example": "veryNicePassword"
                }
              },
              "required": [
                "username",
                "email"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a user",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "adfn123nasdfnj1j129"
                },
                "email": {
                  "type": "string",
                  "example": "lefan@gmail.com"
                },
                "username": {
                  "type": "string",
                  "example": "lefan"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/page/": {
      "get": {
        "tags": [
          "Pages"
        ],
        "summary": "Grab the authenticated user",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a user"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/page/{username}": {
      "get": {
        "tags": [
          "Pages"
        ],
        "summary": "Grab the user",
        "description": "",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the user's page"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/page/pageEntries": {
      "put": {
        "tags": [
          "Pages"
        ],
        "summary": "Update the authenticated user\\'s page.",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "pageEntries": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the updated page"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/media/addProfilePicture": {
      "post": {
        "tags": [
          "Medias"
        ],
        "summary": "Upload a picture to the users folder",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "profilePicture",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "Profile Picture"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/media/{filename}": {
      "delete": {
        "tags": [
          "Medias"
        ],
        "summary": "Delete a file from the user",
        "description": "",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  }
}